requirements.txt
Jinja2==3.1.2
kubernetes
flask
pyyaml
bootstrap-flask
flask-moment
